+++ lib/predicates.dart
@@ -1,3 +1,5 @@
+library predicates;
+
 export 'external/predicates/alternative.dart';
 export 'external/predicates/inverted.dart';
 export 'external/predicates/random.dart';